pyproject.toml
[tool.poetry]
name = "docs-scrap"
version = "0.1.0"
description = ""
authors = ["whatap-docs <kj.im@whatap.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
scrapy = "^2.11.1"
uuid = "^1.30"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

